backend/src/api/get_instability_points_data_api/requirements.txt
fastapi
uvicorn
pytz
numpy
functions-framework==3.*
google-cloud-datastore
gcsfs
zstandard
